pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-atman-env"
version = "1.0.0"
description = "ATMAN: Context-Aware Mobile OS Agent Benchmark"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
]

[project.scripts]
server = "atman_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["atman_env", "atman_env.server"]
package-dir = { "atman_env" = ".", "atman_env.server" = "server" }
